+++ CHANGELOG.md
@@ -11,10 +11,12 @@ CHANGELOG
 #### Landing page
 
 #### Configuration
+- Underscore is now allowed for new parameter names (#153).
 
 #### Management
 
 #### Monitor
+* Fix link to monitor page (#157).
 
 #### Check